B_UtilitiesEndpoint/requirements.txt
boto3==1.26.156
botocore==1.29.156
ffmpeg_python==0.2.0
Flask>=2.2.2
Flask_Cors==4.0.0
psycopg2==2.9.5
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
speechbrain==0.5.16

